move tests to subdirectories
